Remove JarFile and ZipFile from whitelist, add others
While the docs mention that the close method on these classes free
InputStreams created by .getInputStream, a look at the source reveals
that the ctors also open a file and release it on close.

Also took a look at the Eclipse resource leak functionality, and found a
very similar whitelist to ours!